Remove "prepare" stage because it brings confusion for users
Reasoning about prepare stage:
Initially, it was assumed that there would be a significant preparation stage
for the conversion process that wouldn't affect Plesk or the system.
The idea behind this preparation stage was that user start a long preparation,
sure it finished without any problem and schedule the actual conversion,
which would require a long reboot and other related tasks.
However, it was discovered that the preparation stage could be completed
within two minutes, while the whole conversion process took about 30-40 minutes.

Additionally, using control arguments such as '--prepare' and '--start'
increased the risk of making mistakes. As a result, it was decided to remove
these arguments altogether.

So we decided to remove --prepare to avoid confusion from users side.
Especially the stage don't brings much profit.