Workaround to prevent the elevate isscf parser from crashing by adding specific comments
The elevate isscf configuration parser would crash if the named config includes an empty file.
This issue likely occurs because it first attempts to get a character from the string by index
before checking if the index is lower than end_index.
As a solution, we added any commentary string into the included named configuration.
Another option would be to patch the elevate source code, but maintaining the patch can be complex,
so we opted for the workaround until it is fixed on the elevate side.