Move kernel version checker to the KernelVersion class
We should employ KernelVersion to address the issue when the
size of the build part from a version string differs.
Since we anticipate that all missing version parts will be
interpreted as 0, a larger build string with same numbers
should now be recognized as the latest.
However, prior to this change, we were comparing the
build number with the architecture part, which was obviously incorrect.

So now, the build part will be compared with the build part,
and the latest version should be determined correctly.