Transfer centos2alma to the new submodule called dist-upgrader framework
The framework is a component of the distupgrade tools project that
is used to upgrade Ubuntu/Debian family distributions.
It was initially built on the distro-conversion-base, but was forked
to incorporate more core logic and some of actions we use.
We are relocating centos2alma to the framework to streamline maintenance
of both projects and facilitate the sharing of fixes between them.