Change mapping for python-webtest and python-webob packages to python3 packages only
Leapp is expected to install both Python 2 and Python 3 versions
of the webtest and webob modules during the OS conversion.
However, python2-webob is not available in any default repository on AlmaLinux 8,
causing the installation of python2-webtest (which depends on it) to fail as well.

To resolve this, the package mapping in the Leapp configuration
has been adjusted to prevent the installation of unavailable packages.

The workaround introduced in this commit can be disabled
once the missing packages are provided by AlmaLinux 8.