Remove outdated extras repository on AlmaLinux 8 side to prevent conflicts
A Docker extension install a custom extras repo with a randomly generated filename.
In most cases, this doesn't interfere with system behavior.
However, if the filename is alphabetically ordered before `almalinux.repo`,
it overrides the default extras repository, leading to conflicts during
the final stage of the conversion process.

This change adds an action to detect and remove such outdated or potentially conflicting
extras repositories to ensure a smooth conversion.